% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_topics.R
\name{align_topics}
\alias{align_topics}
\title{Align topics from distinct LDA models}
\usage{
align_topics(model_list, models_to_compare)
}
\arguments{
\item{model_list}{(required) a \code{lda_models} object. See \code{run_lda_models} for details.}

\item{models_to_compare}{(optional) either a character indicating if topics should be aligned between \code{consecutive} or \code{all} models, or a list of model pairs between which topics should be aligned.}
}
\value{
a \code{data.frame} (? or some specific object) providing the weights between every pair of topics of each model pairs in the input edgelist (\code{models_to_compare}). ? Do we also return the lda_models with ordered topics?
}
\description{
[Description]
}
