% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_number_of_key_topics.R
\name{compute_number_of_key_topics}
\alias{compute_number_of_key_topics}
\title{Computes the number of key topics}
\usage{
compute_number_of_key_topics(weights, plot = FALSE)
}
\arguments{
\item{weights}{(required) \code{data.frame} with the alignment weights
(the @weight field from an \code{alignment} object)}

\item{plot}{(optional, default = FALSE)
whether to visualize the number of key topics.}
}
\value{
a \code{data.frame} with the number of key topic for each model.
}
\description{
[Description]
}
\seealso{
align_topics
}
