% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_number_of_paths.R
\name{compute_number_of_paths}
\alias{compute_number_of_paths}
\title{Computes the number of paths}
\usage{
compute_number_of_paths(aligned_topics, plot = FALSE)
}
\arguments{
\item{aligned_topics}{(required) An \code{alignment} class with the
alignment results.}

\item{plot}{(optional, default = FALSE) whether to visualize the number of
paths.}
}
\value{
a \code{data.frame} with the number of paths for each model. The
\code{n_topics} column shows the total possible number of topics available
for that model while \code{n_paths} gives the number of paths in that model.
}
\description{
This function computes the number of topics recursively in a backwards
fashion. At the final level of the alignment, the path of each topic is the
topic index. At intermediate levels, paths are built following the edges
with the highest weights from children topics to parent topics.
}
\details{
Given an \code{alignment} object, this function computes the number of
paths at each level in the alignment. The result is a \code{data.frame}
mapping each level to a number from 1 to the number of current levels.
}
\examples{
library(purrr)
data <- rmultinom(10, 20, rep(0.1, 20))
lda_params <- setNames(map(1:5, ~ list(k = .)), 1:5)
lda_models <- run_lda_models(data, lda_params)
alignment <- align_topics(lda_models)
compute_number_of_paths(alignment, plot = TRUE)

}
\seealso{
align_topics
}
