% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_refinement_scores.R
\name{compute_refinement_scores}
\alias{compute_refinement_scores}
\title{Computes the refinement score of each topic in each model}
\usage{
compute_refinement_scores(weights)
}
\arguments{
\item{weights}{(required) \code{data.frame} with the alignment weights
(the @weight field from an \code{alignment} object)}
}
\value{
a \code{data.frame}
with the refinement score of each topic in each model.
}
\description{
This function computes a refinement score for each topic in an alignment. The
refinement score is defined recursively from the leaves. All leaves are given
scores of 1. To find the score of an intermediate node, a weighted average is
taken of the refinement scores for all descendant nodes, with weights coming
from the normalized alignment weights.
}
\examples{
library(purrr)
data <- rmultinom(10, 20, rep(0.1, 20))
lda_params <- setNames(map(1:5, ~ list(k = .)), 1:5)
lda_models <- run_lda_models(data, lda_params)
alignment <- align_topics(lda_models)
compute_refinement_scores(weights(alignment))

}
\seealso{
align_topics align_branches
}
