% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_stability.R
\name{compute_stability_along_branches}
\alias{compute_stability_along_branches}
\title{Computes stability score along branches}
\usage{
compute_stability_along_branches(weights)
}
\arguments{
\item{weights}{(required) \code{data.frame} with the alignment weights
(the @weight field from an \code{alignment} object)}
}
\value{
a \code{data.frame}
with the stability score of each branch along k's.
}
\description{
[Description]
}
\seealso{
align_topics align_branches
}
