% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_lda_models.R
\name{.check_params}
\alias{.check_params}
\title{Check validity of arguments arguments to run_lda_models}
\usage{
.check_params(lda_varying_params_lists, lda_fixed_params_list)
}
\arguments{
\item{lda_varying_params_lists}{(required) a \code{list} specifying the
parameter for each models that needs to be ran. Currently, supported
parameters are "k" (the number of topic), "method" ("VEM" or "Gibbs"), and
"control", a list of type \code{LDAcontrol}. See \code{topicmodels::LDA} for
details and below for examples.}

\item{lda_fixed_params_list}{(optional) a \code{list} specifying the
parameters common to all models to be fitted. Values provided by
\code{lda_fixed_params_list} are overwritten by those provided by
\code{lda_varying_params_lists}.}
}
\description{
For each element of the lda_varying_params_lists, we check the varying params
and we add the fixed params to the varying params list. The end result is a
list of lists. Each sub-list has the element k, method and control.
}
\keyword{internal}
