% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_branches.R
\name{identify_branches}
\alias{identify_branches}
\title{Identifies topic branches along the topic alignment graph}
\usage{
identify_branches(weights)
}
\arguments{
\item{weights}{(required) \code{data.frame} with the alignment weights
(the @weight field from an \code{alignment} object)}
}
\value{
a \code{data.frame}
identifying the branch of each topic in each model.
}
\description{
[Description]
}
\seealso{
align_topics
}
