% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alignment.R
\name{kl_mat}
\alias{kl_mat}
\title{Pairwise KL divergences}
\usage{
kl_mat(p)
}
\arguments{
\item{p}{A matrix whose rows correspond to discrete probability vectors.}
}
\value{
K A matrix whose \code{ij} entry is the KL divergence between rows
\code{i} and \code{j} of \code{p}.
}
\description{
Pairwise KL divergences
}
\keyword{internal}
