% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_topics.R
\name{n_models,alignment-method}
\alias{n_models,alignment-method}
\title{Number of Models Method for Alignment Class}
\usage{
\S4method{n_models}{alignment}(x)
}
\description{
Number of Models Method for Alignment Class
}
