% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_topics.R
\name{n_models}
\alias{n_models}
\title{Generic - Number of Topics}
\usage{
n_models(x)
}
\arguments{
\item{x}{An alignment object output from \code{align_topics}.}
}
\description{
Generic - Number of Topics
}
