% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_topics.R
\name{n_topics,alignment-method}
\alias{n_topics,alignment-method}
\title{Number of Topics Method for Alignment Class}
\usage{
\S4method{n_topics}{alignment}(x)
}
\arguments{
\item{x}{An alignment object output from \code{align_topics}.}
}
\description{
Number of Topics Method for Alignment Class
}
