% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder.R
\name{optimize_permutation}
\alias{optimize_permutation}
\title{Permutation Optimization}
\usage{
optimize_permutation(weights, top_n = 2)
}
\description{
This tries all the permutations and picks one with the lowest cost.
}
