% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alignment.R
\name{plot_alignment}
\alias{plot_alignment}
\title{Visualize alignments between topics of distinct LDA models}
\usage{
plot_alignment(
  aligned_topics,
  add_leaves = TRUE,
  min_beta = 0.025,
  n_features = NULL,
  add_feature_labels = TRUE,
  reverse_x_axis = FALSE
)
}
\arguments{
\item{aligned_topics}{(required)}

\item{add_leaves}{(optional, default = TRUE) whether topic composition of last model should be displayed.}

\item{min_beta}{(optional, default = 0.025) if \code{add_leaves} is \code{TRUE}, this option specifies the minimum beta value (i.e. proportion in topic) for a feature to be displayed in the leaves.}

\item{n_features}{(optional) alternative to \code{min_beta}. if \code{add_leaves} is \code{TRUE}, this option specifies the minimum beta value (i.e. proportion in topic) for a feature to be displayed in the leaves.}

\item{add_feature_labels}{(optional, default = TRUE) if \code{add_leaves} is \code{TRUE}, this option specifies if the name of the features should be displayed.}

\item{reverse_x_axis}{(optional, default = TRUE) specifies whether the x-axis (models) should be reversed.}
}
\value{
a \code{ggplot} object (?) .
}
\description{
[Description]
}
\seealso{
align_topics
}
