% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_number_of_key_topics.R
\name{plot_number_of_key_topics}
\alias{plot_number_of_key_topics}
\title{Plots the number of key topics}
\usage{
plot_number_of_key_topics(n_key_topics)
}
\description{
This function plots the number of key topics *vs.* the number of
topics in each model.
}
