% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_number_of_paths.R
\name{plot_number_of_paths}
\alias{plot_number_of_paths}
\title{Plots the number of paths}
\usage{
plot_number_of_paths(n_paths)
}
\arguments{
\item{n_paths}{A \code{data.frame} with columns \code{n_topics}
and \code{n_key_topics}, as generated by \code{compute_number_of_paths}}
}
\description{
This function plots the number of paths *vs.* the number of
topics in each model.
}
