% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_topics.R
\name{product_weights}
\alias{product_weights}
\title{Product Weights between a Model Pair}
\usage{
product_weights(gammas, ...)
}
\arguments{
\item{gammas}{(required) A list of length two, containing the mixed
membership matrices (a \code{matrix} of dimension n-samples by k-topics) to
compare. The number of columns may be different, but the number of samples
must be equal.}

\item{...}{(optional) Other keyword arguments. These are unused by the
\code{product_weights} alignment strategy, but is included for consistency
across weight functions.}
}
\value{
products A \code{data.frame} giving the product similarity of each
pair of topics across the two input matrices.
}
\description{
An alignment based on product weights sets the weight between topics k and k'
according to \eqn{\gamma_{k}^T\gamma_{k}^\prime}, where \eqn{\gamma_{k} \in
\mathbb{R}^n_{+}} provides the mixed membership assigned to topic \eqn{k}
across the \eqn{n} samples (and similarly for topic \eqn{k^\prime}). This
function computes these weights given a list of two \eqn{n \times K} gamma
matrices.
}
\examples{
g1 <- matrix(runif(20 * 2), 20, 2)
g2 <- matrix(runif(20 * 4), 20, 4)
product_weights(list(g1, g2))

}
\seealso{
align_graph
}
