% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_topics.R
\name{setup_edges}
\alias{setup_edges}
\title{Edgelists for Default Alignments}
\usage{
setup_edges(comparisons, model_names)
}
\arguments{
\item{comparisons}{A string describing the type of model comparisons to
compute.}

\item{model_names}{The names of the models to compare. The resulting edge
list will refer to models by these names.}
}
\description{
This is a helper function for setting up edges that can be used by
align_graph. It implements two types of comparisons, 'consecutive' and 'all'.
It returns a data frame specifying which topics to compare from across all
models.
}
