% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_topics.R
\name{show,alignment-method}
\alias{show,alignment-method}
\title{Show Method for Alignment Class}
\usage{
\S4method{show}{alignment}(object)
}
\arguments{
\item{object}{An alignment object output from \code{align_topics}.}
}
\description{
Show Method for Alignment Class
}
