% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{topic_coherence}
\alias{topic_coherence}
\title{Computes the coherence score of a topic}
\usage{
topic_coherence(model, topic, aligned_topics)
}
\description{
The coherence score of a topic reflects how similar this topic is
to the other topics of the same branch (same key topic).
}
