% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_topics.R
\name{topics,alignment-method}
\alias{topics,alignment-method}
\title{Extract List of Topics and their Summaries}
\usage{
\S4method{topics}{alignment}(x)
}
\arguments{
\item{x}{An alignment object output from \code{align_topics}.}
}
\description{
Extract List of Topics and their Summaries
}
