% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_topics.R
\name{topics}
\alias{topics}
\title{Generic - List of Topics and their Summaries}
\usage{
topics(x)
}
\arguments{
\item{x}{An alignment object output from \code{align_topics}.}
}
\description{
Generic - List of Topics and their Summaries
}
