% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_topics.R
\name{transport_weights}
\alias{transport_weights}
\title{Transport Weights between Model Pair}
\usage{
transport_weights(gammas, betas, reg = 0.1, ...)
}
\arguments{
\item{gammas}{(required) A list of length two, containing the mixed
membership matrices (a \code{matrix} of dimension n-samples by k-topics) to
compare. The number of columns may be different, but the number of samples
must be equal.}
}
\value{
products A \code{data.frame} giving the product similarity of each
pair of topics across the two input matrices.
}
\description{
An alignment based on transport weights sets the weight between topics k and
k' according to an optimal transport problem with (1) costs set by the
distance (specifically, Jensen-Shannon Divergence) between \eqn{\beta_{k}}
and \eqn{\beta_{k^\prime}} and (2) masses defined by the total topic mixed
memberships \eqn{\sum_{i}\gamma_{ik}} and \eqn{\sum_{i}\gamma_{ik^\prime}}.
If topics have similar mixed membership weight and similar topic \eqn{\beta},
then they will be given high transport alignment weight.
}
\examples{
library(purrr)
data <- rmultinom(10, 20, rep(0.1, 20))
lda_params <- setNames(map(1:5, ~ list(k = .)), 1:5)
lda_models <- run_lda_models(data, lda_params)
gammas <- list(lda_models[[1]]$gamma, lda_models[[2]]$gamma)
betas <- list(lda_models[[1]]$beta, lda_models[[2]]$beta)
transport_weights(gammas, betas)

}
