% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_refinement_scores.R
\name{compute_refinement_scores}
\alias{compute_refinement_scores}
\title{Computes the refinement score of each topic in each model}
\usage{
compute_refinement_scores(weights)
}
\arguments{
\item{weights}{(required) \code{data.frame} with the alignment weights
(the @weight field from an \code{alignment} object)}
}
\value{
a \code{data.frame}
with the refinement score of each topic in each model.
}
\description{
[Description]
}
\seealso{
align_topics align_branches
}
