% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_stability.R
\name{compute_stability_along_branches}
\alias{compute_stability_along_branches}
\title{Computes stability score along branches}
\usage{
compute_stability_along_branches(weights)
}
\arguments{
\item{weights}{(required) \code{data.frame} with the alignment weights
(the @weight field from an \code{alignment} object)}
}
\value{
a \code{data.frame} with the stability score of each branch along
k's.
}
\description{
Stability is defined as the product of weights along a branch. Therefore, if
weights are concentrated on specific descendants, then a branch is stable.
This metric is useful for identifying topics that are recovered in models
across a range of values of K (they are "stable" to changes in the K
parameter).
}
\examples{
library(purrr)
data <- rmultinom(10, 20, rep(0.1, 20))
lda_params <- setNames(map(1:5, ~ list(k = .)), 1:5)
lda_models <- run_lda_models(data, lda_params)
alignment <- align_topics(lda_models)
compute_stability_along_branches(weights(alignment))

}
\seealso{
align_topics align_branches
}
