% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alignment.R
\name{discrepancy}
\alias{discrepancy}
\title{Discrepancy for ordering words in plot_betas}
\usage{
discrepancy(p, lambda = 1e-07)
}
\description{
This implements equation (4) from
https://doi.org/10.1371/journal.pgen.1006599
}
