% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder.R
\name{optimize_permutation}
\alias{optimize_permutation}
\title{Naive Optimization}
\usage{
optimize_permutation(weights)
}
\description{
This just tries all the permutations and picks one with the lowest cost.
}
