% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alignment.R
\name{plot,alignment,missing-method}
\alias{plot,alignment,missing-method}
\title{Plot Method for Alignment Class}
\usage{
\S4method{plot}{alignment,missing}(x, reverse_x_axis = FALSE, rect_gap = 0.2, color_by = "branch")
}
\description{
Plot Method for Alignment Class
}
