% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_topics.R
\name{plot,alignment,missing-method}
\alias{plot,alignment,missing-method}
\title{Plot Method for Alignment Class}
\usage{
\S4method{plot}{alignment,missing}(x, y, ...)
}
\description{
Plot Method for Alignment Class
}
