% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alignment.R
\name{plot,alignment,missing-method}
\alias{plot,alignment,missing-method}
\title{Plot Method for Alignment Class}
\usage{
\S4method{plot}{alignment,missing}(
  x,
  add_leaves = TRUE,
  min_beta = 0.025,
  n_features = NULL,
  add_feature_labels = TRUE,
  reverse_x_axis = FALSE,
  rect_hwidth = 0.2
)
}
\description{
Plot Method for Alignment Class
}
