% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alignment.R
\name{plot_alignment}
\alias{plot_alignment}
\title{Visualize alignments between topics of distinct LDA models}
\usage{
plot_alignment(
  x,
  rect_gap = 0.2,
  color_by = "branch",
  model_name_repair_fun = paste0
)
}
\arguments{
\item{x}{(required) An alignment class object resulting from
\code{align_topics}.}

\item{rect_gap}{(optional) A float describing how much vertical space to put
between topics within the same model. The units correspond to topic masses.
Defaults to 0.2.}

\item{color_by}{(optional) What should the color of topics and branches
encode? Defaults to 'branch'. Other possible arguments are 'coherence',
'refinement', or 'topic'.}

\item{model_name_repair_fun}{How should names be repaired before plotting?}
}
\value{
A \code{ggplot2} object describing the alignment weights across
models.
}
\description{
This function creates flow diagrams of alignment weight across collections of
topics. It is the default plot method associated with alignment objects. The
x-axis indexes models from those with few to those with many topics. Each
rectangle matches a topic; its height is the weight associated wtih that
topic from across all samples. The width of edges between topics corresponds
ot the alignment weight between that pair of topics. By default, topics are
shaded according to their branch membership. Other topic measures can be
provided by modifying the `color_by` argument.
}
\seealso{
align_topics
}
