% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alignment.R
\name{plot_alignment}
\alias{plot_alignment}
\title{Visualize alignments between topics of distinct LDA models}
\usage{
plot_alignment(x, reverse_x_axis = FALSE, rect_gap = 0.2, color_by = "branch")
}
\arguments{
\item{reverse_x_axis}{(optional, default = TRUE) specifies whether
the x-axis (models) should be reversed.}

\item{aligned_topics}{(required)}

\item{add_leaves}{(optional, default = TRUE) whether topic composition of
last model should be displayed.}
}
\value{
a \code{ggplot} object (?) .
}
\description{
[Description]
}
\seealso{
align_topics
}
