% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alignment.R
\name{plot_beta}
\alias{plot_beta}
\title{Plot Words Heatmap}
\usage{
plot_beta(x, models = "all", min_beta = 0.025, n_features = NULL, ...)
}
\arguments{
\item{min_beta}{(optional, default = 0.025) if \code{add_leaves} is
\code{TRUE}, this option specifies the minimum beta value (i.e. proportion in
topic) for a feature to be displayed in the leaves.}

\item{n_features}{(optional) alternative to \code{min_beta}. if
\code{add_leaves} is \code{TRUE}, this option specifies the minimum beta
value (i.e. proportion in topic) for a feature to be displayed in the leaves.}

\item{add_feature_labels}{(optional, default = TRUE) if \code{add_leaves} is
\code{TRUE}, this option specifies if the name of the features should be
displayed.}
}
\description{
Plot Words Heatmap
}
