% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_lda_models.R
\name{run_lda_models}
\alias{run_lda_models}
\title{Run LDA models for specified values of hyper-parameters.}
\usage{
run_lda_models(
  data,
  lda_varying_params_lists,
  lda_fixed_params_list = list(),
  dir = NULL,
  reset = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{data}{(required) a \code{matrix}, \code{data.frame} or
\code{slam::simple_triplet_matrix} containing the counts (integers) of each
feature (e.g. words) and each sample (or document). If data is provided as
\code{matrix} or \code{data.frame}, each row is a sample, each column is a
feature.}

\item{lda_varying_params_lists}{(required) a \code{list} specifying the
parameter for each models that needs to be ran. Currently, supported
parameters are "k" (the number of topic), "method" ("VEM" or "Gibbs"), and
"control", a list of type \code{LDAcontrol}. See \code{topicmodels::LDA} for
details and below for examples.}

\item{lda_fixed_params_list}{(optional) a \code{list} specifying the
parameters common to all models to be fitted. Values provided by
\code{lda_fixed_params_list} are overwritten by those provided by
\code{lda_varying_params_lists}.}

\item{dir}{(optional) a \code{character} specifying the directory in which
individual LDA models should be stored. If not specified, individual LDA
models are not stored. This option is especially useful for data exploration
as it allows to save execution time if one wishes to add models to an
existing model list. (see examples)}

\item{reset}{(optional, default = \code{FALSE}). Should any cached models in
the save directory be cleared?}

\item{verbose}{(optional, default = \code{FALSE}) Print verbose output while
running models?}
}
\value{
a list of LDA models (see package \code{topicmodels}).
? or a \code{lda_models} object which would be a list of
1. a list of model;
2. some metadata about the alignement
}
\description{
This function fits a collection of lda models to a dataset, fitting one model
for each hyperparameter setting specified by the
\code{lda_varying_params_list} argument. Its output can be directly used by
\code{align_topics}.
}
\examples{
set.seed(1)
data = matrix(sample(0:1000, size = 24), 4, 6)
lda_varying_params_lists = list(K2 = list(k = 2), K3 = list(k = 3))
lda_models =
   run_lda_models(
      data = data,
      lda_varying_params_lists = lda_varying_params_lists,
      dir = "test_lda_models/"
      )

additional_lda_varying_params_list =
   list(K4 = list(k = 4))
 updated_lda_models =
   run_lda_models(
      data = data,
      lda_varying_params_lists =
        append(
           lda_varying_params_lists,
           additional_lda_varying_params_list),
      dir = "test_lda_models/"
      )

# because we specified the "dir" option, it only runs LDA for k = 4
unlink("test_lda_models/", recursive = TRUE)
}
