% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_lda_models.R
\name{run_lda_models}
\alias{run_lda_models}
\title{Run LDA models for specified values of hyper-parameters or for different environments.}
\usage{
run_lda_models(data, varying, values, controls, reset = FALSE)
}
\arguments{
\item{data}{(required) a \code{matrix} or \code{data.frame} containing the counts (integers) of each feature (e.g. words) and each sample (or document). Each row is a sample, each column is a feature.}

\item{varying}{(required) a \code{character} specifying the varying element across models. Currently supported values for \code{varying} are (i) \code{K}, the number of topics, (ii) \code{alpha}, the prior on the number of topics per document, and (iii) \code{environment}, the environment in which samples have been collected.}

\item{values}{(required)}

\item{controls}{(optional)}

\item{reset}{(optional, default = \code{FALSE})}
}
\value{
a list of LDA models (see package \code{topicmodels}).
}
\description{
[Description]
}
