% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{topic_refinement}
\alias{topic_refinement}
\title{Computes the refinement score of a topic}
\usage{
topic_refinement(model, topic, aligned_topics)
}
\arguments{
\item{model}{The index of the model containing the topic whose coherence we
want}

\item{topic}{The index of the topic whose coherence we want, within the
current model}

\item{aligned_topics}{An alignment object on which the coherence scores
should be computed.}
}
\description{
The refinement score of a topic reflects the average weight
that each of this topic's children attributes to that topic.
A topic with a high refinement score is a topic for which
all of the subsequent topics connected by some forward weight to
this topic have high backward weights to that topic.
}
\keyword{internal}
