% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alignment.R
\name{trim_betas}
\alias{trim_betas}
\title{Filter to words of interest}
\usage{
trim_betas(betas, filter_by, threshold = 0, n_features = NULL)
}
\arguments{
\item{betas}{A \code{data.frame} whose rows correspond to topics and whose
columns are \code{m}, the model ID, and the words for which estimates are
made.}

\item{filter_by}{a character specifying if the data (\code{betas}) should
be filtered by the average \code{"beta"} across topics or
by the \code{"distinctiveness"} of the features.}

\item{threshold}{Words (features) with less than this average beta or
distinctiveness across all topics are ignored}

\item{n_features}{Maximum number of features to show}
}
\description{
Filter to words of interest
}
