% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alignment.R
\name{trim_betas}
\alias{trim_betas}
\title{Filter to words of interest}
\usage{
trim_betas(betas, min_beta = -1, n_features = NULL)
}
\arguments{
\item{min_beta}{Words with less than this beta in all topics are ignored}

\item{n_features}{Maximum number of features to show}
}
\description{
Filter to words of interest
}
